Remove separate hydro headers and add viscosity
